Last Update
Last Update

+++ src/utilities/Xls_Reader.java
@@ -83,7 +83,7 @@ public class Xls_Reader {
 				return cell.getStringCellValue();
 			else if (cell.getCellType() == Cell.CELL_TYPE_NUMERIC || cell.getCellType() == Cell.CELL_TYPE_FORMULA) {
 
-				String cellText = String.valueOf(cell.getNumericCellValue());
+				String cellText = String.valueOf((long)cell.getNumericCellValue());
 				if (HSSFDateUtil.isCellDateFormatted(cell)) {
 
 					double d = cell.getNumericCellValue();
